// fast-crm/src/response-bot/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartMemory, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AGENT_MEMORY: SmartMemory;
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CRM_DATABASE: SmartSql;
  CRM_SERVICE: ServiceStub<import('../crm-service/index.js').default>;
  logger: Logger;
  tracer: Tracer;
  TRIAGE_BOT: ServiceStub<import('../triage-bot/index.js').default>;
}